requirements.txt
pandas
singlestoredb
pymysql
openai>=1.2.3
numpy
python-dotenv
pyautogen